Rename the module address
